requirements.txt
appdirs==1.4.4
attrs==22.1.0
autopep8==2.0.2
beautifulsoup4==4.11.1
bs4==0.0.1
cattrs==22.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
gTTS==2.3.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
numpy==1.24.2
Pillow==9.3.0
playsound==1.3.0
PyAudio==0.2.12
pycodestyle==2.10.0
pygame==2.1.2
pyglet==2.0.0
pylet==0.0.1
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt6==6.4.2
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
python-vlc==3.0.18121
pyttsx3==2.90
pywikihow==0.5.7
regex==2022.10.31
requests==2.28.1
requests-cache==0.9.7
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
tomli==2.0.1
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.65.0
typing_extensions==4.5.0
url-normalize==1.4.3
urllib3==1.26.13
wikipedia==1.4.0
